requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
nltk>=3.8.1
matplotlib>=3.7.0
numpy>=1.24.0
